fin-agent/requirements.txt
crewai
langchain
langchain-openai
python-dotenv
crewai[tools]
pypdf2
streamlit
flask
flask_cors
